Move ElementsService.js into services/
